+++ core/src/com/haniel/Shooter/entities/asteroids/Asteroid128.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.math.Rectangle;
 
 public class Asteroid128 extends Asteroid{
 	
-	public Asteroid128(float x, float y, float moveX, float moveY) {
+	public Asteroid128(double x, double y, float moveX, float moveY) {
 		super(x, y, moveX, moveY);
 		this.xOffset = 6;
 		this.yOffset = 6;
@@ -12,7 +12,7 @@ public class Asteroid128 extends Asteroid{
 		this.height = 114;
 		this.health = width / 8;
 		this.texture = asteroid128_1;
-		this.rectangle = new Rectangle(x + xOffset, y + yOffset, width, height);
+		this.rectangle = new Rectangle((float)x + xOffset,(float) y + yOffset, width, height);
 		
 	}
