+++ core/src/com/haniel/Shooter/entities/asteroids/Asteroid32.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.math.Rectangle;
 
 public class Asteroid32 extends Asteroid{
 	
-	public Asteroid32(float x, float y, float moveX, float moveY) {
+	public Asteroid32(double x, double y, float moveX, float moveY) {
 		super(x, y, moveX, moveY);
 		this.xOffset = 4;
 		this.yOffset = 4;
@@ -12,7 +12,7 @@ public class Asteroid32 extends Asteroid{
 		this.height = 24;
 		this.health = width / 8;
 		this.texture = asteroid32_1;
-		this.rectangle = new Rectangle(x + xOffset, y + yOffset, width, height);
+		this.rectangle = new Rectangle((float)x + xOffset,(float) y + yOffset, width, height);
 		
 	}
